Extract tier normalization to separate file
Move the normalizeTiers implementation out of internalRpc.ts into a new internalRpcTiers.ts as normalizeProviderTiers, and update internalRpc.ts to import and use the new function. Update tests to directly exercise normalizeProviderTiers (removing the previous heavy mocks) and verify bigint amountCents are converted to numbers and undefined input returns an empty array. Behavior is preserved while improving modularity and testability.
