+++ apps/bot/test/lib/internalRpc.test.ts
@@ -1,67 +1,11 @@
-import { afterAll, beforeEach, describe, expect, it, mock } from 'bun:test';
+import { describe, expect, it } from 'bun:test';
 
-const listProviderTiersMock = mock(async () => ({
-  tiers: [
-    {
-      id: 'tier_1',
-      productId: 'campaign_1',
-      name: 'VIP',
-      description: 'Top tier',
-      amountCents: 1500n,
-      currency: 'USD',
-      active: true,
-    },
-  ],
-}));
-
-mock.module('@tempojs/client', () => ({
-  BearerCredential: {
-    create: () => ({
-      storeCredential: mock(async () => undefined),
-    }),
-  },
-  NoStorageStrategy: class NoStorageStrategy {},
-  TempoChannel: {
-    forAddress: () => ({
-      getClient: () => ({
-        listProviderTiers: listProviderTiersMock,
-      }),
-    }),
-  },
-}));
-
-mock.module('@tempojs/common', () => ({
-  ConsoleLogger: class ConsoleLogger {},
-  TempoLogLevel: {
-    Warn: 'warn',
-  },
-}));
-
-mock.module('@yucp/private-rpc', () => ({
-  CatalogClient: class CatalogClient {},
-  CollaboratorClient: class CollaboratorClient {},
-  SetupClient: class SetupClient {},
-  VerificationClient: class VerificationClient {},
-}));
-
-mock.module('@yucp/shared', () => ({
-  getInternalRpcSharedSecret: () => 'shared-secret',
-}));
-
-mock.module('../../src/lib/apiUrls', () => ({
-  getApiUrls: () => ({
-    apiInternal: 'http://127.0.0.1:8787',
-    apiPublic: undefined,
-  }),
-}));
-
-const { listProviderTiers } = await import('../../src/lib/internalRpc');
+const { normalizeProviderTiers } = await import('../../src/lib/internalRpcTiers');
 
 describe('bot internalRpc tier normalization', () => {
-  beforeEach(() => {
-    listProviderTiersMock.mockReset();
-    listProviderTiersMock.mockResolvedValue({
-      tiers: [
+  it('converts bigint tier amounts into numbers for bot product flows', () => {
+    expect(
+      normalizeProviderTiers([
         {
           id: 'tier_1',
           productId: 'campaign_1',
@@ -71,28 +15,19 @@ describe('bot internalRpc tier normalization', () => {
           currency: 'USD',
           active: true,
         },
-      ],
-    });
-  });
+      ])
+    ).toEqual([
+      {
+        id: 'tier_1',
+        productId: 'campaign_1',
+        name: 'VIP',
+        description: 'Top tier',
+        amountCents: 1500,
+        currency: 'USD',
+        active: true,
+      },
+    ]);
 
-  afterAll(() => {
-    mock.restore();
-  });
-
-  it('converts bigint tier amounts into numbers for bot product flows', async () => {
-    await expect(listProviderTiers('patreon', 'creator-user', 'campaign_1')).resolves.toEqual({
-      tiers: [
-        {
-          id: 'tier_1',
-          productId: 'campaign_1',
-          name: 'VIP',
-          description: 'Top tier',
-          amountCents: 1500,
-          currency: 'USD',
-          active: true,
-        },
-      ],
-      error: undefined,
-    });
+    expect(normalizeProviderTiers(undefined)).toEqual([]);
   });
 });
